fix improper emulating of going to the next search result page
Previously, Bing Pong Helper clicked on the "2" on the search result
pages to move to page 2 of the search results. Bing has updated their
search page since then that broke this. Consequently, I have updated
this implementation to click on the next page arrow instead, and to
emulate this click better with a mouse hover and down before clicking.