emulate clicking on each dashboard task (partially implemented)
Instead of taking the list of dashboard tasks, opening them all in new
tabs at once, and then closing them all at once, click on one, wait a
bit, go back to the list of tasks, and then proceed to the next task.
This should satisfy Microsoft's GLinkPing requirements on the dashboard.

NOTE: This feature is only partially working right now. It produces
errors when attempting to click on each task, and it will not return
control to Bing Pong.