more work done on emulating dashboard task clicks
This is an extension of the work done in commit
de5cf1d38ac056db04825a407c14190e52fae041.

This new implementation still doesn't work 100% properly yet. The only
issue right now is that Bing Pong Helper is not doing all of the tasks
it should. This will get ironed out later.