remove problematic login input checking
This is an attempt to fix a bug that causes Bing Pong to hang while
signing into an account. Previously, Bing Pong Helper performed a small
amount of input checking to see if it correctly inputted the account's
username and password in the sign-in tab. For reasons I have forgotten,
if it input checking fails BPH does nothing to attempt to solve this
issue (in an earlier version, BPH similar reattempted inputting the
username/password). This input check is not necessary, and it causes
problems under some circumstances. In particular, these two
circumstances have been causing issues:

(1) Attempting to sign into a Microsoft account that isn’t signed up for
Bing Rewards.

(2) The sign-in succeeds but the account fails Bing Pong’s sign-in
checks due to long Bing Rewards dashboard loading times. When Bing Pong
Helper attempts to open up the sign-in page to preform another sign-in,
the sign-in page redirects to a Microsoft account information page.

In either case, Bing Pong Helper attempts to enter in sign-in details
and fails, causing the hang. Note that Bing Pong also does sign-in
checking, so this input validation is not needed in these two cases.