wait a bit before performing the first dashboard task
This commit fixes the bug referenced here:

https://www.reddit.com/r/bingpong/comments/42zuab/bug_all_dashboard_tasks_not_completed/

Cause: Bing Pong Helper opens up the Bing Rewards dashboard to perform
the dashboard tasks as expected, but it immediately attempts to click on
the last dashboard task before the dashboard even loads, causing the
click to fail. This commit fixes this issue by waiting a certain amount
of time (10 seconds), after which the dashboard has likely loaded.