wait until the dashboard loads before starting timer
Previously, Bing Pong Helper opened the Bing Rewards dashboard and
waited 10 seconds before clicking on the first dashboard task. This is
potentially problematic if the dashboard takes a long time to load.

This commit makes it so that the 10-second timer starts when the
dashboard loads, instead of starting it when the dashboard is opened.