shut up, meg (part 1)
Since some of the dashboard tasks like to have auto-playing audio, their
tabs are now muted.

Also, the dashboard tasks steal focus when they are opened. I have tried
to make this a little less annoying by automatically minimizing the
window when this happens. The tasks will still steal focus for a split
second, but it's an improvement. I am working on improving this.