update version number to v1.3.17.1
I accidentally uploaded the wrong version of this extension to the
Chrome web store, which included the incomplete "Automatically run Bing
Pong daily at ..." option. I have pushed an update out to the store that
removes this incomplete option, and this commit is to get the version
numbers to match again.