another round of dashboard task fixes
This commit fixes several bugs related to performing the dashboard
tasks.

There are two notable bug fixes. The first fix is to implement a maximum
load time for the loading of each task. For some reason, some of the
dashboard tasks tabs won't ever obtain a status of "complete," despite
its tab loading completely. To fix this, if a status of "complete" isn't
obtained within a reasonable period of time, the task will be marked as
done by Bing Pong Helper and it'll move to the next task.

Another bug fixed by this commit is that the last dashboard task in the
list doesn't get completed. This occurred because even though we sent
correctly sent a request to get the task clicked (via line 517), the
"if" statement below it (in the previous commit) would evaluate the
"else" block and immediately close the window. I moved the "if"
statement to another location to fix this.

A third "bug" (although it was more an oversight) is that some of the
dashboard tasks are "naughty." For example, today there is a Twitter
dashboard task that opens up the Microsoft app store on Windows 8/8.1/10
when clicked. There also is another task today that will open up the
Microsoft Edge browser. These tasks are annoying and can interrupt Bing
Pong's normal operation. As a result, I have made these tasks redirect
to a "dashboard task loading blocked" page.